Add option to switch between cyclical and dihedral nodes
This commit is larger than I would have liked, but I wanted the feature
to be complete before I added it. I abstracted the code for interacting
with nodes and for displaying their values, and created new classes for
the graphics part of a vertex and the mathematical group part. Also
added ways in the UI to use them.

This update breaks the legend and the saving and loading. These will be
fixed soon.